Add better variable handling
Variables are now replaced in the regex and the response, with a brand
new syntax: ${variable}. This syntax is much more powerful, allowing for
dynamic variable names using the familiar syntax of responses. For
example, ${variable[A,B]} will be variableA or variableB with a 50/50
chance. Variables that do not exist become the empty string.