Make side effects "real-time" (plus code cleanup)
The biggest change in this commit is implementing generators for
get_messages and friends, which allows the state of a BotBot bot to be
updated between two messages. Previously, if a BotBot bot had this code:

^!test -> {!nick NewNick, My nick is ${self}.}

The output would be:

[NewNick] My nick is OldNick.

This is because the messages were parsed and generated prior to being
sent, so any side effects were not accounted for. Now, the same code
will behave as expected, outputting:

[NewNick] My nick is NewNick.

This, of course, also applies to other user-created variables, which had
the same issue.

In rewriting the code to use generators, I've also cleaned up the code a
little, making it more readable.