Rework directory structure
First, venvs are cool, so add them to the .gitignore.
Second, if I do python3 setup.py develop, it doesn't work properly
because of the package_dirs entry (see pypa/setuptools#230). So, rename
the source directory to botbot so as to not need package_dirs (uglier
in my opinion, but whatever).
Third, use relative imports, because relative imports are cool (and
because DRY).